w1

+++ src/main/java/lits/com/Lesson9/model/Person.java
@@ -11,14 +11,12 @@ import javax.persistence.*;
 public class Person {
 
     @Id@GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private Integer age;
     private Boolean dead;
 
-    public Person() {
-
-    }
+    public Person(){}
 
 
 }
